split runner file: splash, title + credit
