+++ CMakeLists.txt
@@ -115,6 +115,9 @@ set(SOURCES
 	sources/Component/IntroAnimation/IntroAnimationComponent.cpp
 	sources/System/IntroAnimation/IntroAnimationSystem.hpp
 	sources/System/IntroAnimation/IntroAnimationSystem.cpp
+	sources/Runner/SplashScreenScene.cpp
+	sources/Runner/TitleScreenScene.cpp
+	sources/Runner/CreditScene.cpp
 )
 add_executable(bomberman
                sources/main.cpp
